pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "homebase_calendar_sync"
version = "0.1.17"
authors = [
    { name = "David Midlo", email = "dmidlo@gmail.com" }
]
description = "A simple web scraper that reads gethomebase.com's schedule and updates Google Calendar."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "anyio==4.4.0",
    "beautifulsoup4==4.12.3",
    "cachetools==5.3.3",
    "certifi==2024.6.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "cryptography==42.0.7",
    "google-api-core==2.19.0",
    "google-api-python-client==2.131.0",
    "google-auth==2.29.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "googleapis-common-protos==1.63.0",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httplib2==0.22.0",
    "httpx==0.27.0",
    "idna==3.7",
    "inquirerpy==0.3.4",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    "oauthlib==3.2.2",
    "pendulum==3.0.0",
    "pfzy==0.3.4",
    "prompt_toolkit==3.0.45",
    "proto-plus==1.23.0",
    "protobuf==4.25.3",
    "pyasn1==0.6.0",
    "pyasn1_modules==0.4.0",
    "pycparser==2.22",
    "Pygments==2.18.0",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.1",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rich==13.7.1",
    "rsa==4.9",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.5",
    "time-machine==2.14.1",
    "tzdata==2024.1",
    "uritemplate==4.1.1",
    "urllib3==2.2.1",
    "wcwidth==0.2.13"
]
[project.optional-dependencies]
dev = [
    "black==24.4.2",
    "fgv==3.4.0",
    "lick==8.1.7",
    "olorama==0.4.6",
    "istlib==0.3.8",
    "ocutils==0.21.2",
    "ilelock==3.14.0",
    "dentify==2.5.36",
    "mportlib_metadata==7.1.0",
    "jaraco.classes==3.4.0",
    "jaraco.context==5.3.0",
    "jaraco.functools==4.0.1",
    "keyring==25.2.1",
    "mando==0.7.1",
    "more-itertools==10.2.0",
    "mypy-extensions==1.0.0",
    "nh3==0.2.17",
    "nodeenv==1.9.0",
    "packaging==24.0",
    "pathspec==0.12.1",
    "pkginfo==1.11.0",
    "platformdirs==4.2.2",
    "pre-commit==3.7.1",
    "radon==6.0.1",
    "readme_renderer==43.0",
    "requests-toolbelt==1.0.0",
    "rfc3986==2.0.0",
    "ruff==0.4.7",
    "setuptools==70.0.0",
    "twine==5.1.0",
    "virtualenv==20.26.2",
    "wheel==0.43.0",
    "zipp==3.19.1",
]

[project.scripts]
homebase_calendar_sync = "homebase_calendar_sync.homebase_calendar_sync:main"
